refactor(spec): extract validate schema-shape and split temporal tier-2 tests
Move tier-1 layer/plot schema walks into validate-schema-shape with a shared
GEOM_BRANCHES lookup used by structural eligibility. Co-locate temporal
decision-reuse and position-scale data checks into separate test files.
Public validate() behavior is unchanged.
