+++ .changeset/spec-validate-schema-shape-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Extract tier-1 schema shape walks into a focused module with shared GEOM_BRANCHES, and co-locate temporal tier-2 tests with decision-reuse vs position-scale production modules. Public validate() behavior is unchanged.
